rag-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.12
langchain-text-splitters==0.2.2
chromadb==0.5.11
tiktoken==0.7.0

